import legacy version of website
This repository contains the legacy version of FairOddsCalc. Since the tool has become part of a larger project, I’ve decided to remove the old repositories and keep this one as an archive of the original standalone version.

The new version will be fully integrated into the upcoming project and this old site will no longer be maintained or publicly available.
Stay tuned for updates.